service_trend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

# Google API client for YouTube Data API v3
google-api-python-client==2.137.0

# Ortam değişkenleri için
python-dotenv==1.0.1
